Final version if direct bricking.
